[rename] #24 프로젝트 도메인으로 이동
