+++ src/main/java/org/umc/peerre/domain/project/repository/CommentRepository.java
@@ -1,7 +1,7 @@
-package org.umc.peerre.domain.feedback.repository;
+package org.umc.peerre.domain.project.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.umc.peerre.domain.feedback.entity.Comment;
+import org.umc.peerre.domain.project.entity.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment,Long> {
