+++ src/main/java/org/umc/peerre/domain/user/entity/User.java
@@ -2,7 +2,7 @@ package org.umc.peerre.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.umc.peerre.domain.feedback.entity.Comment;
+import org.umc.peerre.domain.project.entity.Comment;
 import org.umc.peerre.domain.feedback.entity.FeedbackAggregation;
 import org.umc.peerre.domain.feedback.entity.FeedbackRegistration;
 import org.umc.peerre.domain.teamspace.entity.UserTeamspace;
